Added searching include directories; implemented defined() preprocessor operator.
Now, #include "" and #include <> are searched in lists of directories specified in the variables:
vacietis:*include-quote-directories* and vacietis:*include-bracket-directories*; the "current
directory" is searched unless vacietis:*include-disable-current-directory* is true.
The search is performed in a way similar to gcc with its -I, -I- -iquote options.
gcc -nostdinc  can be implemented by removing :libc from *include-bracket-directories*
and gcc -I- by moving :libc in the *include-bracket-directories* list.